Arreglo del seeder, migration de Zona_Dependencia_Programa

+++ database/migrations/2022_11_04_224506_create_zona__dependencia__programas_table.php
@@ -15,12 +15,10 @@ return new class extends Migration
     {
         Schema::create('zona__dependencia__programas', function (Blueprint $table) {
             $table->id();
-            //$table->unsignedBigInteger('user_id');
-            //$table->foreign('user_id')->references('id')->on('users');
-            
-            $table->unsignedBigInteger('id_zona') ; 
-            $table->unsignedBigInteger('clave_dependencia') ; 
-            $table->unsignedBigInteger('clave_programa') ; 
+            /* php artisan migrate --path=database\migrations\2022_11_04_224506_create_zona__dependencia__programas_table.php */
+            $table->unsignedBigInteger('id_zona') ;
+            $table->integer('clave_dependencia') ;
+            $table->integer('clave_programa') ;
 
             $table->foreign('id_zona')->references('id')->on('zonas');
             $table->foreign('clave_dependencia')->references('clave')->on('dependencias');
